altered JSON class hierarchy, its now a lot cleaner

+++ Telematics.Server.Data/DataModels/VehicleSpeed.cs
@@ -19,7 +19,7 @@ namespace Telematics.Server.Data.DataModels
         public int VehicleID { get; set; }
         public Nullable<double> Lat { get; set; }
         public Nullable<double> Lon { get; set; }
-        public Nullable<decimal> Speed { get; set; }
+        public Nullable<long> Speed { get; set; }
         public byte[] EventTime { get; set; }
     
         public virtual User User { get; set; }
